requirements.txt
streamlit==1.32.0
gspread==6.0.0
google-auth==2.28.0
PyMuPDF==1.23.26
easyocr==1.7.1
pdfrw==0.4